+++ src/arena/components.rs
@@ -1,7 +1,15 @@
 use specs::prelude::*;
 use specs_derive::Component;
 
+mod animation;
+pub use animation::{tick_animations, AnimationComponent};
+
+use sdl2::pixels::Color;
+
 use crate::clash::Character;
+
+#[derive(Component)]
+pub struct PlayerComponent {}
 
 #[derive(Hash, PartialEq, Eq, Component)]
 pub struct PositionComponent {
@@ -16,9 +24,6 @@ impl PositionComponent {
 }
 
 #[derive(Component)]
-pub struct PlayerComponent {}
-
-#[derive(Component)]
 pub struct CharacterInfoComponent {
     pub character: Character,
 }
@@ -26,5 +31,18 @@ pub struct CharacterInfoComponent {
 impl CharacterInfoComponent {
     pub const fn init(character: Character) -> CharacterInfoComponent {
         CharacterInfoComponent { character }
+    }
+}
+
+#[derive(Component)]
+pub struct FieldComponent {
+    pub color: Color,
+}
+
+impl FieldComponent {
+    pub fn init(r: u8, g: u8, b: u8) -> FieldComponent {
+        FieldComponent {
+            color: Color::from((r, g, b, 140)),
+        }
     }
 }
